Refactor Auth0 related functions.
