+++ common/models/customer-data.js
@@ -2,48 +2,13 @@
 const appRoot = require('app-root-path');
 const Promise = require('bluebird');
 const _ = require('lodash');
-const jwksClient = require('jwks-rsa');
-const jwtNode = require('jsonwebtoken');
 const objectHash = require('object-hash');
 const app = require(appRoot + '/server/server');
 const logger = require(appRoot + '/config/winston');
+const Auth0Helper = require(appRoot + '/common/util/auth0-helper');
 
 module.exports = function(CustomerData) {
-  const APP_METADATA_KEY = 'https://om.com/app_metadata';
   const CLIENT_MODELS = ['Order', 'Statement', 'EndUser', 'Client'];
-  const client = jwksClient({
-    cache: true,
-    rateLimit: true,
-    jwksUri: `https://${process.env.AUTH0_DOMAIN_ID}/.well-known/jwks.json`
-  });
-
-  function getKey(header, callback) {
-    client.getSigningKey(header.kid, function(err, key) {
-      if (err) {
-        callback(err);
-      } else {
-        let signingKey = key.publicKey || key.rasPublicKey;
-        callback(null, signingKey);
-      }
-    });
-  }
-
-  async function decodeIdToken(idToken) {
-    try {
-      return await new Promise((resolve, reject) => {
-        jwtNode.verify(idToken, getKey, { algorithms: ['RS256'] }, function(err, decoded) {
-          if (err) {
-            reject(err);
-          } else {
-            resolve(decoded);
-          }
-        });
-      });
-    } catch (error) {
-      logger.error(`Failed to verify idToken - ${error.message}`);
-      throwAuthError();
-    }
-  }
 
   function throwAuthError() {
     let error = new Error('User not authenticated');
@@ -66,8 +31,8 @@ module.exports = function(CustomerData) {
    */
   async function verifyIdToken(idToken, modelName, methodName) {
     try {
-      let decoded = await decodeIdToken(idToken);
-      const role = app.models.EndUser.getHighestRole(_.get(decoded, [APP_METADATA_KEY, 'roles'], []));
+      let decoded = await Auth0Helper.decodeToken(idToken);
+      const role = app.models.EndUser.getHighestRole(Auth0Helper.getMetadata(decoded, 'roles', []));
       if ((modelName === 'EndUser' || modelName === 'Client')) {
         // EndUser and Client requires additional access checking
         if (!_.includes(app.models[modelName].allowedMethods(role), methodName)) {
@@ -89,7 +54,7 @@ module.exports = function(CustomerData) {
       }
       if (_.includes(CLIENT_MODELS, modelName)) {
         // limit scope to the user's client
-        _.set(filter, ['where', 'clientId'], _.get(decoded, [APP_METADATA_KEY, 'clientId']));
+        _.set(filter, ['where', 'clientId'], Auth0Helper.getMetadata(decoded, 'clientId'));
       }
 
       return await app.models[modelName].find(filter);
@@ -107,7 +72,7 @@ module.exports = function(CustomerData) {
       }
       if (_.includes(CLIENT_MODELS, modelName)) {
         // limit scope to the user's client
-        _.set(filter, ['where', 'clientId'], _.get(decoded, [APP_METADATA_KEY, 'clientId']));
+        _.set(filter, ['where', 'clientId'], Auth0Helper.getMetadata(decoded, 'clientId'));
       }
       return await app.models[modelName].findById(id, filter);
     } catch (error) {
@@ -141,7 +106,8 @@ module.exports = function(CustomerData) {
           throwAuthError();
         }
         // Client method must have 'clientId' as the first argument.
-        if (modelName === 'Client' && _.get(decoded, [APP_METADATA_KEY, 'clientId']) !== _.get(params, '0')) {
+        if (modelName === 'Client' &&
+          Auth0Helper.getMetadata(decoded, 'clientId') !== _.get(params, '0')) {
           throwAuthError();
         }
       }
